requirements.txt
Django==4.2.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
uwsgi==2.0.26
psycopg==3.2.1